Externalize map data loading

+++ src/util.rs
@@ -13,10 +13,10 @@ impl IsometricOperations for Vec2 {
     const MAP_TILE_HEIGHT: f32 = 64.0;
 
     fn to_isometric(&self) -> Self {
-        let isometric_x = (self.y() * Self::MAP_TILE_WIDTH / 2.0)
-            + (self.x() as f32 * Self::MAP_TILE_WIDTH / 2.0);
-        let isometric_y = (self.x() as f32 * Self::MAP_TILE_HEIGHT / 2.0)
-            - (self.y() as f32 * Self::MAP_TILE_HEIGHT / 2.0);
+        let isometric_x =
+            (self.y() * Self::MAP_TILE_WIDTH / 2.0) + (self.x() * Self::MAP_TILE_WIDTH / 2.0);
+        let isometric_y =
+            (self.x() * Self::MAP_TILE_HEIGHT / 2.0) - (self.y() * Self::MAP_TILE_HEIGHT / 2.0);
 
         Self::new(isometric_x, isometric_y)
     }
